chore: remove dead code and rename stuff
