--- packages/screens/OrganizerDeployer/multisig-proposal-transaction.json
@@ -1,132 +1,0 @@
-[
-    {
-      "id": 1,
-      "type": "transfer",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    },
-    {
-      "id": 2,
-      "type": "stake",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    },
-    {
-      "id": 3,
-      "type": "transfer",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    },
-    {
-      "id": 4,
-      "type": "transfer",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    },
-    {
-      "id": 5,
-      "type": "stake",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    },
-    {
-      "id": 6,
-      "type": "transfer",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    },
-    {
-      "id": 7,
-      "type": "stake",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    },
-    {
-      "id": 8,
-      "type": "transfer",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    },
-    {
-      "id": 9,
-      "type": "stake",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    },
-    {
-      "id": 10,
-      "type": "transfer",
-      "createdAt": "12/9/2022",
-      "time": "10:25PM",
-      "sending": "TORI",
-      "amount": "50",
-      "networkFee": "0.005346",
-      "createdBy": "username",
-      "approvers": 3,
-      "approvedBy": 1,
-      "approvedRequired": 2
-    }
-]
